requirements.txt
ruff
crewai
crewai-tools
langchain
langchain-openai
langchain-community
pydantic
python-dotenv
requests
beautifulsoup4
praw
google-api-python-client